Adds temporary nouislider hack - reassigns value of `document`
The slider handles weren't responding to manipulation.
The original material uses the regular document for almost
everything when it should be using the target's document instead.
It meant the events for the sliders were not being detected.
This hacks the distribution file, not the src files to dangerously
replace the root `document` var value with our own iframe value.