Completely reworks playback internals for more flexibility
The method of traversing the data is still under consideration.
It is done by a sort of false index atm. It may be more clear to
blatantly walk around the arrays, but the main source of confusion
could end up being the incrementation object that gets passed
around. We'll have to see how this holds up.