Switches sentence parsing module, fixes sentence navigation
Before, traveling with sentence navigation included new line
characters and would skip sentences that weren't supposed to be
skipped. This was because sbd was putting new line characters
at the beginning of sentences. The npl_compromise module puts
them at the end of a sentence. Still not ideal because in future
we want to be able to show new paragraph symbols and be able to
see them during sentence navigation, but it's still better than
what we had before.